pyproject.toml
[tool.poetry]
name = "youtube-channel-to-gpt"
version = "0.1.0"
description = "api for youtube project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.11.14"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.9.0"
attrs = "25.3.0"
bcrypt = "4.3.0"
cachetools = "5.5.2"
certifi = "2025.1.31"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
click = "8.1.8"
cryptography = "44.0.2"
defusedxml = "0.7.1"
deprecation = "2.1.0"
distro = "1.9.0"
dnspython = "2.7.0"
ecdsa = "0.19.1"
email-validator = "2.2.0"
fastapi = "0.115.12"
fastapi-cli = "0.0.7"
frozenlist = "1.5.0"
google-api-core = "2.24.2"
google-api-python-client = "2.166.0"
google-auth = "2.38.0"
google-auth-httplib2 = "0.2.0"
googleapis-common-protos = "1.69.2"
gotrue = "2.12.0"
h11 = "0.14.0"
h2 = "4.2.0"
hpack = "4.1.0"
httpcore = "1.0.7"
httplib2 = "0.22.0"
httptools = "0.6.4"
httpx = "0.28.1"
hyperframe = "6.1.0"
idna = "3.10"
jinja2 = "3.1.6"
jiter = "0.9.0"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
mdurl = "0.1.2"
multidict = "6.2.0"
openai = "1.69.0"
passlib = "1.7.4"
pluggy = "1.5.0"
postgrest = "1.0.1"
propcache = "0.3.1"
proto-plus = "1.26.1"
protobuf = "6.30.2"
psycopg2-binary = "^2.9.10"
pyasn1 = "0.4.8"
pyasn1-modules = "0.4.1"
pycparser = "2.22"
pydantic = "2.11.0"
pydantic-core = "2.33.0"
pyjwt = "2.10.1"
pyparsing = "3.2.3"
pytest = "8.3.5"
pytest-mock = "3.14.0"
python-dotenv = "1.1.0"
python-jose = "3.4.0"
python-multipart = "0.0.20"
pyyaml = "6.0.2"
realtime = "2.4.2"
requests = "2.32.3"
rich = "13.9.4"
rich-toolkit = "0.14.0"
rsa = "4.9"
setuptools = "75.8.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.46.1"
storage3 = "0.11.3"
strenum = "0.4.15"
stripe = "11.6.0"
supabase = "2.15.0"
supafunc = "0.9.4"
tqdm = "4.67.1"
typer = "0.15.2"
typing-inspection = "0.4.0"
uritemplate = "4.1.1"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
watchfiles = "1.0.4"
websockets = "14.2"
wheel = "0.45.1"
yarl = "1.18.3"
youtube-transcript-api = "1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
